output/vcd: use larger data type to internally store frequency
The 'period' member of the VCD output module's context is supposed to
hold frequencies that correspond to the timescale used during export.
An 'int' (in combination with VCD's 1/10/100 constraint) thus would
result in a 1GHz limit, use uint64_t instead to support higher rates.

+++ src/output/vcd.c
@@ -31,7 +31,7 @@ struct context {
 	int num_enabled_channels;
 	uint8_t *prevsample;
 	gboolean header_done;
-	int period;
+	uint64_t period;
 	int *channel_index;
 	uint64_t samplerate;
 	uint64_t samplecount;
